refactor(skills): split lib load/registry and trust policy orchestration (#1688)
