+++ crates/tau-skills/src/commands.rs
@@ -1,9 +1,9 @@
 use std::{
-    collections::HashMap,
+    collections::{HashMap, HashSet},
     path::{Path, PathBuf},
 };
 
-use anyhow::{bail, Result};
+use anyhow::{anyhow, bail, Context, Result};
 use serde::Serialize;
 use tau_core::{current_unix_timestamp, is_expired_unix};
